refactor: basic refactor
