+++ src/cpcli/commands/__init__.py
@@ -7,9 +7,9 @@ from zope.interface import Interface, implementer
 from zope.interface.exceptions import Invalid, MultipleInvalid
 from zope.interface.verify import verifyClass
 
-from cpcli.cli import Scraper
-from cpcli.utils.cmdtypes import readable_file, contest_uri
 from cpcli.utils.config import CpCliConfig
+from cpcli.helpers.scrapper import Scraper
+from cpcli.utils.cmdtypes import readable_file, contest_uri
 from cpcli.utils.constants import CONTEST_URI_HELP
 from cpcli.utils.misc import walk_modules
