+++ src/cpcli/commands/show.py
@@ -2,8 +2,8 @@ from argparse import ArgumentParser, Namespace
 
 from zope.interface import implementer
 
-from cpcli.cli import Scraper
 from cpcli.commands import ICommand
+from cpcli.helpers.scrapper import Scraper
 
 
 @implementer(ICommand)
