+++ src/cpcli/commands/testcase.py
@@ -3,7 +3,7 @@ from argparse import ArgumentParser, Namespace
 from zope.interface import implementer
 
 from cpcli.commands import ICommand
-from cpcli.cli import Scraper
+from cpcli.helpers.scrapper import Scraper
 from cpcli.utils.python import multiline_input
